backend/requirements.txt
fastapi
uvicorn[standard]
tinydb
pydantic
python-multipart
pypdf
pillow
python-docx
pdf2image
docx2pdf
pywin32
python-magic-bin
aiofiles
